Back-End/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pymongo>=4.10.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pyjwt>=2.0.0
passlib[argon2]>=1.7.4
jinja2>=3.1.4
markupsafe>=3.0.2
weasyprint>=63.0
